+++ generators/typescript/sdk/cli/src/custom-config/SdkCustomConfig.ts
@@ -88,6 +88,7 @@ export interface SdkCustomConfig {
     resolveQueryParameterNameConflicts: boolean;
     maxRetries: number | undefined;
     alwaysSendAuth: boolean;
+    optionalAuth: boolean;
     retryStatusCodes: "legacy" | "recommended";
     generateReactQueryHooks: boolean;
 }