+++ packages/ir-sdk/src/sdk/serialization/resources/publish/types/PublishTarget.ts
@@ -10,6 +10,7 @@ import { NpmPublishTarget } from "./NpmPublishTarget.js";
 import { NugetPublishTarget } from "./NugetPublishTarget.js";
 import { PostmanPublishTarget } from "./PostmanPublishTarget.js";
 import { PypiPublishTarget } from "./PypiPublishTarget.js";
+import { RubyGemsPublishTarget } from "./RubyGemsPublishTarget.js";
 
 export const PublishTarget: core.serialization.Schema<serializers.PublishTarget.Raw, FernIr.PublishTarget> =
     core.serialization
@@ -21,6 +22,7 @@ export const PublishTarget: core.serialization.Schema<serializers.PublishTarget.
             crates: CratesPublishTarget,
             go: GoPublishTarget,
             nuget: NugetPublishTarget,
+            rubygems: RubyGemsPublishTarget,
         })
         .transform<FernIr.PublishTarget>({
             transform: (value) => {
@@ -39,6 +41,8 @@ export const PublishTarget: core.serialization.Schema<serializers.PublishTarget.
                         return FernIr.PublishTarget.go(value);
                     case "nuget":
                         return FernIr.PublishTarget.nuget(value);
+                    case "rubygems":
+                        return FernIr.PublishTarget.rubygems(value);
                     default:
                         return value as FernIr.PublishTarget;
                 }
@@ -54,7 +58,8 @@ export declare namespace PublishTarget {
         | PublishTarget.Pypi
         | PublishTarget.Crates
         | PublishTarget.Go
-        | PublishTarget.Nuget;
+        | PublishTarget.Nuget
+        | PublishTarget.Rubygems;
 
     export interface Postman extends PostmanPublishTarget.Raw {
         type: "postman";
@@ -82,5 +87,9 @@ export declare namespace PublishTarget {
 
     export interface Nuget extends NugetPublishTarget.Raw {
         type: "nuget";
+    }
+
+    export interface Rubygems extends RubyGemsPublishTarget.Raw {
+        type: "rubygems";
     }
 }