+++ seed/java-sdk/bytes-upload/src/main/java/com/seed/bytesUpload/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.bytes-upload/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:bytes-upload-sdk");
+                put("X-Fern-SDK-Name", "com.fern:bytes-upload");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;