+++ seed/java-sdk/inferred-auth-implicit/src/main/java/com/seed/inferredAuthImplicit/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.inferred-auth-implicit/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:inferred-auth-implicit-sdk");
+                put("X-Fern-SDK-Name", "com.fern:inferred-auth-implicit");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;