+++ seed/java-sdk/java-global-header-env/src/main/java/com/seed/javaGlobalHeaderEnv/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.java-global-header-env/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:java-global-header-env-sdk");
+                put("X-Fern-SDK-Name", "com.fern:java-global-header-env");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;