+++ seed/java-sdk/java-required-body-optional-headers/src/main/java/com/seed/javaRequiredBodyOptionalHeaders/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.java-required-body-optional-headers/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:java-required-body-optional-headers-sdk");
+                put("X-Fern-SDK-Name", "com.fern:java-required-body-optional-headers");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;