+++ seed/java-sdk/multiple-request-bodies/src/main/java/com/seed/api/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.multiple-request-bodies/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:api-sdk");
+                put("X-Fern-SDK-Name", "com.fern:multiple-request-bodies");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;