+++ seed/java-sdk/pagination-custom/default/src/main/java/com/seed/pagination/core/ClientOptions.java
@@ -49,7 +49,8 @@ public final class ClientOptions {
             {
                 put("User-Agent", "com.fern.pagination-custom/0.0.1");
                 put("X-Fern-Language", "JAVA");
-                put("X-Fern-SDK-Name", "com.seed.fern:pagination-sdk");
+                put("X-Fern-SDK-Name", "com.fern:pagination-custom");
+                put("X-Fern-SDK-Version", "0.0.1");
             }
         });
         this.headerSuppliers = headerSuppliers;